+++ flash/core/classification.py
@@ -15,15 +15,15 @@ from typing import Any, Callable, Dict, List, Mapping, Optional, Sequence, Union
 
 import torch
 import torch.nn.functional as F
-import torchmetrics
 from pytorch_lightning.utilities import rank_zero_deprecation, rank_zero_warn
+from torchmetrics import Accuracy, Metric
 
 from flash.core.adapter import AdapterTask
 from flash.core.data.io.input import DataKeys
 from flash.core.data.io.output import Output
 from flash.core.model import Task
 from flash.core.registry import FlashRegistry
-from flash.core.utilities.imports import _FIFTYONE_AVAILABLE, lazy_import, requires
+from flash.core.utilities.imports import _FIFTYONE_AVAILABLE, _TM_GREATER_EQUAL_0_7_0, lazy_import, requires
 from flash.core.utilities.providers import _FIFTYONE
 
 if _FIFTYONE_AVAILABLE:
@@ -34,6 +34,11 @@ else:
     fol = None
     Classification = None
     Classifications = None
+
+if _TM_GREATER_EQUAL_0_7_0:
+    from torchmetrics import F1Score
+else:
+    from torchmetrics import F1 as F1Score
 
 
 CLASSIFICATION_OUTPUTS = FlashRegistry("outputs")
@@ -50,7 +55,7 @@ class ClassificationMixin:
         num_classes: Optional[int] = None,
         labels: Optional[List[str]] = None,
         loss_fn: Optional[Callable] = None,
-        metrics: Union[torchmetrics.Metric, Mapping, Sequence, None] = None,
+        metrics: Union[Metric, Mapping, Sequence, None] = None,
         multi_label: bool = False,
     ):
         self.num_classes = num_classes
@@ -58,7 +63,7 @@ class ClassificationMixin:
         self.labels = labels
 
         if metrics is None:
-            metrics = torchmetrics.F1(self.num_classes) if multi_label and num_classes else torchmetrics.Accuracy()
+            metrics = F1Score(num_classes) if (multi_label and num_classes) else Accuracy()
 
         if loss_fn is None:
             loss_fn = binary_cross_entropy_with_logits if multi_label else F.cross_entropy
@@ -80,7 +85,7 @@ class ClassificationTask(Task, ClassificationMixin):
         *args,
         num_classes: Optional[int] = None,
         loss_fn: Optional[Callable] = None,
-        metrics: Union[torchmetrics.Metric, Mapping, Sequence, None] = None,
+        metrics: Union[Metric, Mapping, Sequence, None] = None,
         multi_label: bool = False,
         labels: Optional[List[str]] = None,
         **kwargs,
@@ -105,7 +110,7 @@ class ClassificationAdapterTask(AdapterTask, ClassificationMixin):
         *args,
         num_classes: Optional[int] = None,
         loss_fn: Optional[Callable] = None,
-        metrics: Union[torchmetrics.Metric, Mapping, Sequence, None] = None,
+        metrics: Union[Metric, Mapping, Sequence, None] = None,
         multi_label: bool = False,
         labels: Optional[List[str]] = None,
         **kwargs,